抽出 VndbPanel 的格式化函数到 utils/vndbFormat.ts（1139 → 974 行）
八个纯函数整块搬出（原 L846-L1031）：formatDate、formatPlatform、
formatLanguage、renderDescription、formatDevStatus、getDevStatusColor、
getTagCategoryClass、formatRelation。它们输入 VNDB 原始值、输出给用户看的
文案或 Tailwind 类名，无副作用、与渲染无关，留在组件里只是历史原因。

renderedDescription 那个 computed 留在组件里 —— 它读 showOriginal /
translatedDescription 两个响应式状态，属于组件状态而非格式化。

逐条比对过行为，包括几处容易改错的 fallback：formatPlatform 与 formatLanguage
未命中映射表时返回的是 `.toUpperCase()` 而不是原值。renderDescription 是这里
唯一有安全含义的函数（涉及 v-html），转义顺序与 URL 协议白名单原样保留，
实测 [url=javascript:...] 仍被拦截、原始 HTML 仍被转义、https 链接仍正常渲染。
